fix(staff): extract pure helpers so unit tests can run
computeProjectsKpis and listProjectMembersPreview both imported MikroORM entities at module top-level, so the unit tests added for SPEC-069 Step 9 transitively loaded @mikro-orm/core ESM and exploded under the jest preset. The tests never executed.

Move the pure helpers into their own files that don't import entities:

- timesheets-projects/kpiMath.ts: deltaPct, minutesToHours

- timesheets-projects/initials.ts: computeInitials

Update the helper test files to import from the new pure modules. computeProjectsKpis and listProjectMembersPreview now import (and re-export computeInitials) from the new files so callers keep working.

Result: 26/26 helper tests now pass.

Fixes review finding #8 on PR #1111.
